MyPage -> InfoOfSooolList 추가하였습니다. 후에 Generic으로 가능한지 알아봐야겠다.
