+++ app/src/main/java/com/example/taopr/soool/Object/InfoOfSooolList.java
@@ -1,0 +1,17 @@
+package com.example.taopr.soool.Object;
+
+import java.util.ArrayList;
+
+public class InfoOfSooolList {
+
+    private ArrayList<InfoOfSoool> infoOfSoools;
+
+    public ArrayList<InfoOfSoool> getInfoOfSoools() {
+        return infoOfSoools;
+    }
+
+    public void setInfoOfSoools(ArrayList<InfoOfSoool> infoOfSoools) {
+        this.infoOfSoools = infoOfSoools;
+    }
+
+}
